requirements.txt
altair==2.3.0
appnope==0.1.0
asn1crypto==0.24.0
atomicwrites==1.3.0
attrdict==2.0.0
attrs==19.2.0
backcall==0.1.0
boto3==1.9.241
botocore==1.12.241
bravado==10.4.1
bravado-core==5.13.2
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
docutils==0.15.2
-e git+https://github.com/facebookresearch/EGG.git@5b422b41334214ac0d43b21f43bc0988569de469#egg=EGG
entrypoints==0.3
future==0.17.1
gitdb2==2.0.6
GitPython==3.0.3
hyperopt==0.1.1
idna==2.8
importlib-metadata==0.23
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonref==0.2
jsonschema==3.0.2
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
monotonic==1.5
more-itertools==7.2.0
msgpack-python==0.5.6
neptune-client==0.4.4
neptune-contrib==0.9.1
networkx==2.3
numpy==1.17.2
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
packaging==19.2
pandas==0.25.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pluggy==0.13.0
prompt-toolkit==2.0.9
psutil==5.6.6
ptyprocess==0.6.0
py==1.8.0
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pymongo==3.9.0
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.2.0
python-dateutil==2.8.0
python-telegram-bot==11.1.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
retrying==1.3.3
s3transfer==0.2.1
scikit-learn==0.21.3
scikit-optimize==0.5.2
scikit-plot==0.3.7
scipy==1.3.1
seaborn==0.9.0
simplejson==3.16.0
six==1.12.0
smmap2==2.0.5
swagger-spec-validator==2.4.3
tabulate==0.8.3
toolz==0.10.0
torch==1.1.0
torchvision==0.3.0
tqdm==4.35.0
traitlets==4.3.2
typing-extensions==3.7.4
urllib3==1.25.6
vega==2.6.0
wcwidth==0.1.7
websocket-client==0.56.0
zipp==0.6.0
